+++ tests/_config/components.php
@@ -8,7 +8,10 @@ return [
             'article' => [
                 'class' => \vintage\search\tests\models\Article::className(),
                 'label' => 'Test label'
-            ]
+            ],
+            'homeStaticPage' => [
+                'class' => \vintage\search\tests\models\HomeStaticPage::className(),
+            ],
         ]
     ]
 ];