+++ tests/_data/dump.sql
@@ -9,7 +9,15 @@ CREATE TABLE article (
   shortText STRING,
   fullText TEXT,
   socialNetworksText STRING
-  );
+);
+
+-- Table: static_page
+DROP TABLE IF EXISTS static_page;
+CREATE TABLE static_page (
+  id INTEGER PRIMARY KEY AUTOINCREMENT,
+  slug STRING,
+  description TEXT
+);
 
 COMMIT TRANSACTION;
 PRAGMA foreign_keys = on;