+++ tests/_data/home-static-page.php
@@ -1,0 +1,18 @@
+<?php
+return [
+    [
+        'id' => 1,
+        'slug' => 'about',
+        'description' => 'This is about page',
+    ],
+    [
+        'id' => 2,
+        'slug' => 'contacts',
+        'description' => 'This is contacts page',
+    ],
+    [
+        'id' => 3,
+        'slug' => 'home',
+        'description' => 'This is home page',
+    ],
+];