+++ CHANGELOG.md
@@ -3,7 +3,9 @@ Search Component Change Log
 
 1.1.0 Under development
 -----------------------
-* Enh: Refactored code: improve code readability.
+* Enh: Refactored code: improve code readability
+* Enh: Creates interface for custom search query
+* Enh: Improved unit tests
 
 1.0.0 August 11, 2017
 ---------------------